Added bonds portfolio selecting algorithm

+++ FinTrader.Pro.Contracts/Bonds/Portfolio.cs
@@ -6,7 +6,7 @@ namespace FinTrader.Pro.Contracts.Bonds
     public class Portfolio
     {
         public string Includes { get; set; }
-        public int Sum { get; set; }
+        public double Sum { get; set; }
         public int Pay { get; set; }
         public double Yields { get; set; }
         public DateTime MatDate { get; set; }
